rename to ops app name
