+++ .secrets.baseline
@@ -97,7 +97,7 @@
         "verified_result": null
       }
     ],
-    "backend/ops_api/ops_site/fixtures/fake_data.json": [
+    "backend/ops_api/ops/fixtures/fake_data.json": [
       {
         "hashed_secret": "6b8a1c79a05009129a1f92df6b370300adbabd56",
         "is_secret": false,
