+++ backend/deployed_run.sh
@@ -16,7 +16,7 @@ python ./ops_api/manage.py migrate
 
 if [[ "${SPACE_NAME}" == "john.skinner" ]]; then
     # Load the fake data because we're deploying to the dev environment
-    python ./ops_api/manage.py loaddata ./ops_api/ops_site/fixtures/fake_data.json
+    python ./ops_api/manage.py loaddata ./ops_api/ops/fixtures/fake_data.json
 fi
 
 python -m gunicorn ops_api.django_config.asgi:application --worker-class=uvicorn.workers.UvicornWorker --bind=0.0.0.0:8080
