+++ backend/ops_api/ops/portfolios/controller.py
@@ -2,8 +2,8 @@ from rest_framework import serializers
 from rest_framework.generics import ListAPIView
 from rest_framework.generics import RetrieveAPIView
 
-from ops_api.ops_site.cans.controller import CANSerializer
-from ops_api.ops_site.portfolios.models import Portfolio
+from ops_api.ops.cans.controller import CANSerializer
+from ops_api.ops.portfolios.models import Portfolio
 
 
 class PortfolioSerializer(serializers.ModelSerializer):
