+++ backend/tests/ops_site/cans/test_can_controller.py
@@ -1,5 +1,5 @@
-from ops_api.ops_site.cans.controller import CANFiscalYearSerializer
-from ops_api.ops_site.cans.controller import CANSerializer
+from ops_api.ops.cans.controller import CANFiscalYearSerializer
+from ops_api.ops.cans.controller import CANSerializer
 
 
 def test_CAN_serializer_has_depth_of_one():
