+++ docs/models.dot
@@ -22,7 +22,7 @@ digraph model_graph {
   // Labels
 
 
-  ops_api_ops_site_models_Role [label=<
+  ops_api_ops_models_Role [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -48,7 +48,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_models_Person [label=<
+  ops_api_ops_models_Person [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -90,7 +90,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_portfolios_models_Portfolio [label=<
+  ops_api_ops_portfolios_models_Portfolio [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -140,7 +140,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_FundingPartner [label=<
+  ops_api_ops_cans_models_FundingPartner [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -174,7 +174,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_CommonAccountingNumber [label=<
+  ops_api_ops_cans_models_CommonAccountingNumber [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -248,7 +248,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_CANFiscalYear [label=<
+  ops_api_ops_cans_models_CANFiscalYear [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -314,7 +314,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_Contract [label=<
+  ops_api_ops_cans_models_Contract [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -340,7 +340,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_ContractLineItem [label=<
+  ops_api_ops_cans_models_ContractLineItem [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -374,7 +374,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_ContractLineItemFiscalYear [label=<
+  ops_api_ops_cans_models_ContractLineItemFiscalYear [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -408,7 +408,7 @@ digraph model_graph {
     </TABLE>
     >]
 
-  ops_api_ops_site_cans_models_ContractLineItemFiscalYearPerCAN [label=<
+  ops_api_ops_cans_models_ContractLineItemFiscalYearPerCAN [label=<
     <TABLE BGCOLOR="white" BORDER="1" CELLBORDER="0" CELLSPACING="0">
     <TR><TD COLSPAN="2" CELLPADDING="5" ALIGN="CENTER" BGCOLOR="#1b563f">
     <FONT FACE="Roboto" COLOR="white" POINT-SIZE="10"><B>
@@ -455,37 +455,37 @@ digraph model_graph {
 
   // Relations
 
-  ops_api_ops_site_models_Person -> ops_api_ops_site_models_Role
+  ops_api_ops_models_Person -> ops_api_ops_models_Role
   [label=" roles (person)"] [arrowhead=dot arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_CommonAccountingNumber -> ops_api_ops_site_cans_models_FundingPartner
+  ops_api_ops_cans_models_CommonAccountingNumber -> ops_api_ops_cans_models_FundingPartner
   [label=" authorizer (authorizer)"] [arrowhead=none, arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_CommonAccountingNumber -> ops_api_ops_site_portfolios_models_Portfolio
+  ops_api_ops_cans_models_CommonAccountingNumber -> ops_api_ops_portfolios_models_Portfolio
   [label=" portfolio (portfolio)"] [arrowhead=none, arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_CommonAccountingNumber -> ops_api_ops_site_cans_models_FundingPartner
+  ops_api_ops_cans_models_CommonAccountingNumber -> ops_api_ops_cans_models_FundingPartner
   [label=" funding_source (commonaccountingnumber)"] [arrowhead=dot arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_CANFiscalYear -> ops_api_ops_site_cans_models_CommonAccountingNumber
+  ops_api_ops_cans_models_CANFiscalYear -> ops_api_ops_cans_models_CommonAccountingNumber
   [label=" can (canfiscalyear)"] [arrowhead=none, arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_CANFiscalYear -> ops_api_ops_site_models_Person
+  ops_api_ops_cans_models_CANFiscalYear -> ops_api_ops_models_Person
   [label=" can_lead (canfiscalyear)"] [arrowhead=dot arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_Contract -> ops_api_ops_site_cans_models_CommonAccountingNumber
+  ops_api_ops_cans_models_Contract -> ops_api_ops_cans_models_CommonAccountingNumber
   [label=" cans (contracts)"] [arrowhead=dot arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_ContractLineItem -> ops_api_ops_site_cans_models_Contract
+  ops_api_ops_cans_models_ContractLineItem -> ops_api_ops_cans_models_Contract
   [label=" contract (line_items)"] [arrowhead=none, arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_ContractLineItemFiscalYear -> ops_api_ops_site_cans_models_ContractLineItem
+  ops_api_ops_cans_models_ContractLineItemFiscalYear -> ops_api_ops_cans_models_ContractLineItem
   [label=" line_item (fiscal_years)"] [arrowhead=none, arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_ContractLineItemFiscalYearPerCAN -> ops_api_ops_site_cans_models_ContractLineItemFiscalYear
+  ops_api_ops_cans_models_ContractLineItemFiscalYearPerCAN -> ops_api_ops_cans_models_ContractLineItemFiscalYear
   [label=" fiscal_year (cans)"] [arrowhead=none, arrowtail=dot, dir=both];
 
-  ops_api_ops_site_cans_models_ContractLineItemFiscalYearPerCAN -> ops_api_ops_site_cans_models_CommonAccountingNumber
+  ops_api_ops_cans_models_ContractLineItemFiscalYearPerCAN -> ops_api_ops_cans_models_CommonAccountingNumber
   [label=" can (line_items_fy)"] [arrowhead=none, arrowtail=dot, dir=both];
